fix: wallet top-up, auth keyboard overflow, test stabilization
Wallet (6 bugs):
- Add _centavosFromJson converter to WalletAccount.balance and
  WalletTransaction.amount to safely parse Postgres numeric to Dart int
- Fix tx.amount.toStringAsFixed(2) compile error (int has no method)
  -> (tx.amount / 100).toStringAsFixed(2)
- Rewire CashInScreen to actually call topUp() provider (was no-op)
  with loading state and real success/error SnackBar feedback
- Add ref.listen error SnackBar to wallet screen for AsyncError states
- Fix home_screen.dart: replace 'Juan Dela Cruz' placeholder with 'Rider'

Auth / Keyboard (2 screens):
- Fix login_screen: Spacer() caused RenderFlex overflow when soft keyboard
  opens on physical device -> SingleChildScrollView(reverse:true) + Gap
  Lottie hides via AnimatedContainer when keyboard is open
- Fix otp_screen: same Spacer() overflow bug -> same pattern fix

Tests (3 pre-existing failures, now passing):
- home_screen_test: update 'Where to?' -> 'Where to or What to buy?'
- home_screen_test + accessibility_test: mock permission_handler channel
  to prevent MissingPluginException crash
- core_components_golden_test: allow runtime font fetch for NotoSansJP
  (WeeshAppBar uses GoogleFonts.notoSansJp which is not a bundled asset)

All 18 tests pass. flutter analyze: no errors.

+++ weesh_mobile/lib/features/wallet/domain/repositories/wallet_repository.dart
@@ -4,5 +4,5 @@ import 'package:weesh_mobile/features/wallet/domain/models/wallet_transaction.da
 abstract class WalletRepository {
   Future<WalletAccount> getOrCreateWallet(String userId);
   Future<List<WalletTransaction>> getRecentTransactions(String walletId);
-  Future<void> topUpWallet(String walletId, double amount, String source);
+  Future<void> topUpWallet(String walletId, int amount, String source);
 }
